test(core): split watch update tests
