+++ packages/core/test/core/watch/updates/handle-callbacks.ts
@@ -1,0 +1,77 @@
+import {
+  describe,
+  test,
+  expect,
+  watch,
+  mkdtempSync,
+  writeFileSync,
+  rmSync,
+  join,
+  tmpdir,
+} from './helpers';
+
+describe('watch() > rebuild updates > handle and callbacks', () => {
+  test('stop() 후 리빌드 발생하지 않음', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-watch-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 1;');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    let rebuildCount = 0;
+
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      onReady() {
+        readyDone();
+      },
+      onRebuild() {
+        rebuildCount++;
+      },
+    });
+
+    await readyP;
+    handle.stop();
+
+    // stop 후 파일 수정
+    await new Promise((r) => setTimeout(r, 100));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 2;');
+    await new Promise((r) => setTimeout(r, 1000));
+
+    expect(rebuildCount).toBe(0);
+    rmSync(dir, { recursive: true });
+  }, 5000);
+
+  test('double stop()은 에러 없이 무시', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-watch-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 1;');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      onReady() {
+        readyDone();
+      },
+    });
+
+    await readyP;
+    handle.stop();
+    // 두 번째 stop() — 에러 없이 무시되어야 함
+    expect(() => handle.stop()).not.toThrow();
+    rmSync(dir, { recursive: true });
+  });
+
+  test('콜백 없이 watch — crash 없이 동작', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-watch-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 1;');
+
+    // onReady, onRebuild 모두 미제공
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+    });
+
+    // 초기 빌드 완료 대기 (콜백 없으므로 타이머로)
+    await new Promise((r) => setTimeout(r, 1500));
+    expect(() => handle.stop()).not.toThrow();
+    rmSync(dir, { recursive: true });
+  }, 5000);
+});
